+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpec.kt
@@ -24,6 +24,7 @@ import org.jetbrains.intellij.IntelliJPluginConstants.POST_INSTRUMENT_TEST_CODE_
 import org.jetbrains.intellij.IntelliJPluginConstants.PREPARE_SANDBOX_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.PREPARE_TESTING_SANDBOX_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.PREPARE_UI_TESTING_SANDBOX_TASK_NAME
+import org.jetbrains.intellij.IntelliJPluginConstants.PRINT_PRODUCTS_RELEASES_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.PUBLISH_PLUGIN_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.RUN_IDE_FOR_UI_TESTS_TASK_NAME
 import org.jetbrains.intellij.IntelliJPluginConstants.RUN_IDE_PERFORMANCE_TEST_TASK_NAME
@@ -65,6 +66,7 @@ class IntelliJPluginSpec : IntelliJPluginSpecBase() {
                 PREPARE_SANDBOX_TASK_NAME,
                 PREPARE_TESTING_SANDBOX_TASK_NAME,
                 PREPARE_UI_TESTING_SANDBOX_TASK_NAME,
+                PRINT_PRODUCTS_RELEASES_TASK_NAME,
                 PUBLISH_PLUGIN_TASK_NAME,
                 RUN_IDE_TASK_NAME,
                 RUN_IDE_FOR_UI_TESTS_TASK_NAME,